Stop testing key if correct key is found
The Keybits of the status register are updated after every send key, so
just sending all possible key's without stopping after the right one
tends to fail.

this also allows us to check the status before sending any key to see if
a correct key is already provided.
This is very usefull because the bootloader seems to 'recognize' every
key as wrong after the correct key was send.